+++ scout/server/app.py
@@ -13,6 +13,7 @@ from flask import current_app, Flask, redirect, request, url_for
 from flask_babel import Babel
 from flask_login import current_user
 from flaskext.markdown import Markdown
+from scout.utils.matchmaker import mme_nodes
 
 logger = logging.getLogger(__name__)
 
@@ -27,7 +28,7 @@ except ImportError:
     logger.info('chanjo report not installed!')
 
 from . import extensions
-from .blueprints import (alignviewers, public, genes, cases, login, variants, panels, dashboard, 
+from .blueprints import (alignviewers, public, genes, cases, login, variants, panels, dashboard,
                          api, phenotypes, institutes)
 
 
@@ -40,6 +41,11 @@ def create_app(config_file=None, config=None):
         app.config.update(config)
     if config_file:
         app.config.from_pyfile(config_file)
+
+    # If there is a MatchMaker Exchange server
+    # collect the connected external nodes
+    app.mme_nodes = mme_nodes(app.config.get('MME_URL'), app.config.get('MME_TOKEN'))
+
 
     app.config["JSON_SORT_KEYS"] = False
     current_log_level = logger.getEffectiveLevel()