+++ scout/server/blueprints/cases/controllers.py
@@ -8,25 +8,33 @@ import logging
 
 from bs4 import BeautifulSoup
 from xlsxwriter import Workbook
-from flask import url_for
+from flask import url_for, current_app
 from flask_mail import Message
 import query_phenomizer
 from flask_login import current_user
 
-from scout.constants import (CASE_STATUSES, PHENOTYPE_GROUPS, COHORT_TAGS, SEX_MAP, PHENOTYPE_MAP, VERBS_MAP, MT_EXPORT_HEADER)
+from scout.constants import (CASE_STATUSES, PHENOTYPE_GROUPS, COHORT_TAGS, SEX_MAP, PHENOTYPE_MAP,
+                             CANCER_PHENOTYPE_MAP, VERBS_MAP, MT_EXPORT_HEADER)
 from scout.constants.variant_tags import MANUAL_RANK_OPTIONS, DISMISS_VARIANT_OPTIONS, GENETIC_MODELS
 from scout.export.variant import export_mt_variants
 from scout.server.utils import institute_and_case, user_institutes
 from scout.parse.clinvar import clinvar_submission_header, clinvar_submission_lines
 from scout.server.blueprints.variants.controllers import variant as variant_decorator
 from scout.server.blueprints.variants.controllers import sv_variant
+from scout.parse.matchmaker import hpo_terms, omim_terms, genomic_features, parse_matches
+from scout.utils.matchmaker import matchmaker_request
 from scout.server.blueprints.variants.controllers import get_predictions
 from scout.server.blueprints.genes.controllers import gene
 
-log = logging.getLogger(__name__)
+LOG = logging.getLogger(__name__)
+
 
 STATUS_MAP = {'solved': 'bg-success', 'archived': 'bg-warning'}
 
+TRACKS = {
+    'rare': 'Rare Disease',
+    'cancer': 'Cancer',
+}
 
 def cases(store, case_query, limit=100):
     """Preprocess case objects.
@@ -52,6 +60,8 @@ def cases(store, case_query, limit=100):
         case_groups[case_obj['status']].append(case_obj)
         case_obj['is_rerun'] = len(case_obj.get('analyses', [])) > 0
         case_obj['clinvar_variants'] = store.case_to_clinVars(case_obj['_id'])
+        case_obj['display_track'] = TRACKS[case_obj.get('track', 'rare')]
+
     data = {
         'cases': [(status, case_groups[status]) for status in CASE_STATUSES],
         'found_cases': case_query.count(),
@@ -82,7 +92,12 @@ def case(store, institute_obj, case_obj):
         except ValueError as err:
             sex = 0
         individual['sex_human'] = SEX_MAP[sex]
-        individual['phenotype_human'] = PHENOTYPE_MAP.get(individual['phenotype'])
+
+        pheno_map = PHENOTYPE_MAP
+        if case_obj.get('track', 'rare') == 'cancer':
+            pheno_map = CANCER_PHENOTYPE_MAP
+
+        individual['phenotype_human'] = pheno_map.get(individual['phenotype'])
         case_obj['individual_ids'].append(individual['individual_id'])
 
     case_obj['assignees'] = [store.user(user_email) for user_email in
@@ -105,15 +120,17 @@ def case(store, institute_obj, case_obj):
         full_name = "{} ({})".format(panel_obj['display_name'], panel_obj['version'])
         case_obj['panel_names'].append(full_name)
     case_obj['default_genes'] = list(distinct_genes)
-
     for hpo_term in itertools.chain(case_obj.get('phenotype_groups', []),
                                     case_obj.get('phenotype_terms', [])):
         hpo_term['hpo_link'] = ("http://compbio.charite.de/hpoweb/showterm?id={}"
                                 .format(hpo_term['phenotype_id']))
 
     # other collaborators than the owner of the case
-    o_collaborators = [store.institute(collab_id) for collab_id in case_obj['collaborators'] if
-                       collab_id != case_obj['owner']]
+    o_collaborators = []
+    for collab_id in case_obj['collaborators']:
+        if collab_id != case_obj['owner'] and store.institute(collab_id):
+            o_collaborators.append(store.institute(collab_id))
+
     case_obj['o_collaborators'] = [(collab_obj['_id'], collab_obj['display_name']) for
                                    collab_obj in o_collaborators]
 
@@ -126,7 +143,11 @@ def case(store, institute_obj, case_obj):
     for event in events:
         event['verb'] = VERBS_MAP[event['verb']]
 
-    case_obj['clinvar_variants'] = store.case_to_clinVars(case_obj['display_name'])
+    case_obj['clinvar_variants'] = store.case_to_clinVars(case_obj['_id'])
+
+    # if updated_at is a list, set it to the last update datetime
+    if case_obj.get('updated_at') and isinstance(case_obj['updated_at'], list):
+        case_obj['updated_at'] = max(case_obj['updated_at'])
 
     # Phenotype groups can be specific for an institute, there are some default groups
     pheno_groups = institute_obj.get('phenotype_groups') or PHENOTYPE_GROUPS
@@ -141,7 +162,9 @@ def case(store, institute_obj, case_obj):
         'causatives': causatives,
         'collaborators': collab_ids,
         'cohort_tags': COHORT_TAGS,
+        'mme_nodes': current_app.mme_nodes, # Get available MatchMaker nodes for matching case
     }
+
     return data
 
 
@@ -356,7 +379,7 @@ def mt_excel_files(store, case_obj, temp_excel_dir):
 def update_synopsis(store, institute_obj, case_obj, user_obj, new_synopsis):
     """Update synopsis."""
     # create event only if synopsis was actually changed
-    if new_synopsis and case_obj['synopsis'] != new_synopsis:
+    if case_obj['synopsis'] != new_synopsis:
         link = url_for('cases.case', institute_id=institute_obj['_id'],
                        case_name=case_obj['display_name'])
         store.update_synopsis(institute_obj, case_obj, user_obj, link,
@@ -423,6 +446,7 @@ def update_default_panels(store, current_user, institute_id, case_name, panel_id
     panel_objs = [store.panel(panel_id) for panel_id in panel_ids]
     store.update_default_panels(institute_obj, case_obj, user_obj, link, panel_objs)
 
+
 def vcf2cytosure(store, institute_id, case_name, individual_id):
     """vcf2cytosure CGH file for inidividual."""
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
@@ -442,19 +466,41 @@ def gene_variants(store, variants_query, page=1, per_page=50):
 
     my_institutes = list(inst['_id'] for inst in user_institutes(store, current_user))
 
-    log.debug("Institutes allowed: {}.".format(my_institutes))
-
     variants = []
     for variant_obj in variant_res:
         # hide other institutes for now
-        if (variant_obj['institute'] not in my_institutes):
-            log.debug("Institute {} not allowed.".format(variant_obj['institute']))
+        if variant_obj['institute'] not in my_institutes:
+            LOG.warning("Institute {} not allowed.".format(variant_obj['institute']))
             continue
 
+        # Populate variant case_display_name
         variant_case_obj = store.case(case_id=variant_obj['case_id'])
-        case_display_name = variant_case_obj['display_name']
+        if not variant_case_obj:
+            # A variant with missing case was encountered
+            continue
+        case_display_name = variant_case_obj.get('display_name')
         variant_obj['case_display_name'] = case_display_name
 
+        genome_build = variant_case_obj.get('genome_build', '37')
+        if genome_build not in ['37','38']:
+            genome_build = '37'
+
+        # Update the HGNC symbols if they are not set
+        variant_genes = variant_obj.get('genes')
+        if variant_genes is not None:
+            for gene_obj in variant_genes:
+                # If there is no hgnc id there is nothin we can do
+                if not gene_obj['hgnc_id']:
+                    continue
+                # Else we collect the gene object and check the id
+                if gene_obj.get('hgnc_symbol') is None or gene_obj.get('description') is None:
+                    hgnc_gene = store.hgnc_gene(gene_obj['hgnc_id'], build=genome_build)
+                    if not hgnc_gene:
+                        continue
+                    gene_obj['hgnc_symbol'] = hgnc_gene['hgnc_symbol']
+                    gene_obj['description'] = hgnc_gene['description']
+
+        # Populate variant HGVS and predictions
         gene_ids = []
         gene_symbols = []
         hgvs_c = []
@@ -463,6 +509,7 @@ def gene_variants(store, variants_query, page=1, per_page=50):
 
         if variant_genes is not None:
             functional_annotation = ''
+
             for gene_obj in variant_genes:
                 hgnc_id = gene_obj['hgnc_id']
                 gene_symbol = gene(store, hgnc_id)['symbol']
@@ -479,8 +526,6 @@ def gene_variants(store, variants_query, page=1, per_page=50):
                 hgvs_c.append(hgvs_nucleotide)
                 hgvs_p.append(hgvs_protein)
 
-            log.debug("HGVS: {} {} {}.".format(gene_symbols, hgvs_c, hgvs_p))
-
             if len(gene_symbols) == 1:
                 if(hgvs_p[0] != "None"):
                     hgvs = hgvs_p[0]
@@ -489,7 +534,10 @@ def gene_variants(store, variants_query, page=1, per_page=50):
                 else:
                     hgvs = "-"
                 variant_obj['hgvs'] = hgvs
+
+            # populate variant predictions for display
             variant_obj.update(get_predictions(variant_genes))
+
         variants.append(variant_obj)
 
     return {
@@ -562,3 +610,222 @@ def get_sanger_unevaluated(store, institute_id, user_id):
             unevaluated.append(unevaluated_by_case)
 
     return unevaluated
+
+
+def mme_add(store, user_obj, case_obj, add_gender, add_features, add_disorders, genes_only,
+    mme_base_url, mme_accepts, mme_token):
+    """Add a patient to MatchMaker server
+
+    Args:
+        store(adapter.MongoAdapter)
+        user_obj(dict) a scout user object (to be added as matchmaker contact)
+        case_obj(dict) a scout case object
+        add_gender(bool) if True case gender will be included in matchmaker
+        add_features(bool) if True HPO features will be included in matchmaker
+        add_disorders(bool) if True OMIM diagnoses will be included in matchmaker
+        genes_only(bool) if True only genes and not variants will be shared
+        mme_base_url(str) base url of the MME server
+        mme_accepts(str) request content accepted by MME server
+        mme_token(str) auth token of the MME server
+
+    Returns:
+        submitted_info(dict) info submitted to MatchMaker and its responses
+    """
+
+    if not mme_base_url or not mme_accepts or not mme_token:
+        return 'Please check that Matchmaker connection parameters are valid'
+    url = ''.join([mme_base_url, '/patient/add'])
+    features = []   # this is the list of HPO terms
+    disorders = []  # this is the list of OMIM diagnoses
+    g_features = []
+
+     # create contact dictionary
+    contact_info = {
+        'name' : user_obj['name'],
+        'href' : ''.join( ['mailto:',user_obj['email']] ),
+        'institution' : 'Scout software user, Science For Life Laboratory, Stockholm, Sweden'
+    }
+    if add_features: # create features dictionaries
+        features = hpo_terms(case_obj)
+
+    if add_disorders: # create OMIM disorders dictionaries
+        disorders = omim_terms(case_obj)
+
+    # send a POST request and collect response for each affected individual in case
+    server_responses = []
+
+    submitted_info = {
+        'contact' : contact_info,
+        'sex' : add_gender,
+        'features' : features,
+        'disorders' : disorders,
+        'genes_only' : genes_only,
+        'patient_id' : []
+    }
+    for individual in case_obj.get('individuals'):
+        if not individual['phenotype'] in  [2, 'affected']: # include only affected individuals
+            continue
+
+        patient = {
+            'contact' : contact_info,
+            'id' : '.'.join([case_obj['_id'], individual.get('individual_id')]), # This is a required field form MME
+            'label' : '.'.join([case_obj['display_name'], individual.get('display_name')]),
+            'features' : features,
+            'disorders' : disorders
+        }
+        if add_gender:
+            if individual['sex'] == '1':
+                patient['sex'] = 'MALE'
+            else:
+                patient['sex'] = 'FEMALE'
+
+        if case_obj.get('suspects'):
+            g_features = genomic_features(store, case_obj, individual.get('display_name'), genes_only)
+            patient['genomicFeatures'] = g_features
+
+        # send add request to server and capture response
+        resp = matchmaker_request(url=url, token=mme_token, method='POST', content_type=mme_accepts,
+            accept='application/json', data={'patient':patient})
+
+        server_responses.append({
+                'patient': patient,
+                'message': resp.get('message'),
+                'status_code' : resp.get('status_code')
+            })
+    submitted_info['server_responses'] = server_responses
+    return submitted_info
+
+
+def mme_delete(case_obj, mme_base_url, mme_token):
+    """Delete all affected samples for a case from MatchMaker
+
+    Args:
+        case_obj(dict) a scout case object
+        mme_base_url(str) base url of the MME server
+        mme_token(str) auth token of the MME server
+
+    Returns:
+         server_responses(list): a list of object of this type:
+                    {
+                        'patient_id': patient_id
+                        'message': server_message,
+                        'status_code': server_status_code
+                    }
+    """
+    server_responses = []
+
+    if not mme_base_url or not mme_token:
+        return 'Please check that Matchmaker connection parameters are valid'
+
+    # for each patient of the case in matchmaker
+    for patient in case_obj['mme_submission']['patients']:
+
+        # send delete request to server and capture server's response
+        patient_id = patient['id']
+        url = ''.join([mme_base_url, '/patient/delete/', patient_id])
+        resp = matchmaker_request(url=url, token=mme_token, method='DELETE', )
+
+        server_responses.append({
+            'patient_id': patient_id,
+            'message': resp.get('message'),
+            'status_code': resp.get('status_code')
+        })
+
+    return server_responses
+
+
+def mme_matches(case_obj, institute_obj, mme_base_url, mme_token):
+    """Show Matchmaker submission data for a sample and eventual matches.
+
+    Args:
+        case_obj(dict): a scout case object
+        institute_obj(dict): an institute object
+        mme_base_url(str) base url of the MME server
+        mme_token(str) auth token of the MME server
+
+    Returns:
+        data(dict): data to display in the html template
+    """
+    data = {
+        'institute' : institute_obj,
+        'case' : case_obj,
+        'server_errors' : []
+    }
+    matches = {}
+    # loop over the submitted samples and get matches from the MatchMaker server
+    if not case_obj.get('mme_submission'):
+        return None
+
+    for patient in case_obj['mme_submission']['patients']:
+        patient_id = patient['id']
+        matches[patient_id] = None
+        url = ''.join([ mme_base_url, '/matches/', patient_id])
+        server_resp = matchmaker_request(url=url, token=mme_token, method='GET')
+        if 'status_code' in server_resp: # the server returned a valid response
+            # and this will be a list of match objects sorted by desc date
+            pat_matches = []
+            if server_resp.get('matches'):
+                pat_matches = parse_matches(patient_id, server_resp['matches'])
+            matches[patient_id] = pat_matches
+        else:
+            LOG.warning('Server returned error message: {}'.format(server_resp['message']))
+            data['server_errors'].append(server_resp['message'])
+
+    data['matches'] = matches
+
+    return data
+
+
+def mme_match(case_obj, match_type, mme_base_url, mme_token, nodes=None, mme_accepts=None):
+    """Initiate a MatchMaker match against either other Scout patients or external nodes
+
+    Args:
+        case_obj(dict): a scout case object already submitted to MME
+        match_type(str): 'internal' or 'external'
+        mme_base_url(str): base url of the MME server
+        mme_token(str): auth token of the MME server
+        mme_accepts(str): request content accepted by MME server (only for internal matches)
+
+    Returns:
+        matches(list): a list of eventual matches
+    """
+    query_patients = []
+    server_responses = []
+    url = None
+    # list of patient dictionaries is required for internal matching
+    query_patients = case_obj['mme_submission']['patients']
+    if match_type=='internal':
+        url = ''.join([mme_base_url,'/match'])
+        for patient in query_patients:
+            json_resp = matchmaker_request(url=url, token=mme_token, method='POST',
+                content_type=mme_accepts, accept=mme_accepts, data={'patient':patient})
+            resp_obj = {
+                'server' : 'Local MatchMaker node',
+                'patient_id' : patient['id'],
+                'results' : json_resp.get('results'),
+                'status_code' : json_resp.get('status_code'),
+                'message' : json_resp.get('message') # None if request was successful
+            }
+            server_responses.append(resp_obj)
+    else: # external matching
+        # external matching requires only patient ID
+        query_patients = [ patient['id'] for patient in query_patients]
+        node_ids = [ node['id'] for node in nodes ]
+        if match_type in node_ids: # match is against a specific external node
+            node_ids = [match_type]
+
+        # Match every affected patient
+        for patient in query_patients:
+            # Against every node
+            for node in node_ids:
+                url = ''.join([mme_base_url,'/match/external/', patient, '?node=', node])
+                json_resp = matchmaker_request(url=url, token=mme_token, method='POST')
+                resp_obj = {
+                    'server' : node,
+                    'patient_id' : patient,
+                    'results' : json_resp.get('results'),
+                    'status_code' : json_resp.get('status_code'),
+                    'message' : json_resp.get('message') # None if request was successful
+                }
+                server_responses.append(resp_obj)
+    return server_responses